backend/requirements.txt
google-generativeai>=0.8.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
httpx>=0.27.0
apscheduler>=3.10.0
slowapi>=0.1.9

email-validator
pydantic[email]
